+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Additional glossary terms for publication frequency and link types (contribution from @effjot Florian Jenn)
+
 ## 1.2.2 - 2023-05-15
 
 * Fix "array exceeds maximum" for layers with many features (contribution from @effjot Florian Jenn)